Move ShardReplicationStatuses out of wrangler for import
