+++ go/vt/wrangler/reparent.go
@@ -23,71 +23,20 @@ This file handles the reparenting operations.
 import (
 	"context"
 	"fmt"
-	"sync"
 	"time"
 
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	"vitess.io/vitess/go/event"
-	"vitess.io/vitess/go/vt/concurrency"
 	"vitess.io/vitess/go/vt/log"
-	"vitess.io/vitess/go/vt/topo"
 	"vitess.io/vitess/go/vt/topo/topoproto"
-	"vitess.io/vitess/go/vt/topotools"
 	"vitess.io/vitess/go/vt/topotools/events"
 	"vitess.io/vitess/go/vt/vtctl/grpcvtctldserver"
 	"vitess.io/vitess/go/vt/vtctl/reparentutil"
 
-	replicationdatapb "vitess.io/vitess/go/vt/proto/replicationdata"
 	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
 	vtctldatapb "vitess.io/vitess/go/vt/proto/vtctldata"
 )
-
-// ShardReplicationStatuses returns the ReplicationStatus for each tablet in a shard.
-func (wr *Wrangler) ShardReplicationStatuses(ctx context.Context, keyspace, shard string) ([]*topo.TabletInfo, []*replicationdatapb.Status, error) {
-	tabletMap, err := wr.ts.GetTabletMapForShard(ctx, keyspace, shard)
-	if err != nil {
-		return nil, nil, err
-	}
-	tablets := topotools.CopyMapValues(tabletMap, []*topo.TabletInfo{}).([]*topo.TabletInfo)
-
-	wr.logger.Infof("Gathering tablet replication status for: %v", tablets)
-	wg := sync.WaitGroup{}
-	rec := concurrency.AllErrorRecorder{}
-	result := make([]*replicationdatapb.Status, len(tablets))
-
-	for i, ti := range tablets {
-		// Don't scan tablets that won't return something
-		// useful. Otherwise, you'll end up waiting for a timeout.
-		if ti.Type == topodatapb.TabletType_PRIMARY {
-			wg.Add(1)
-			go func(i int, ti *topo.TabletInfo) {
-				defer wg.Done()
-				pos, err := wr.tmc.PrimaryPosition(ctx, ti.Tablet)
-				if err != nil {
-					rec.RecordError(fmt.Errorf("PrimaryPosition(%v) failed: %v", ti.AliasString(), err))
-					return
-				}
-				result[i] = &replicationdatapb.Status{
-					Position: pos,
-				}
-			}(i, ti)
-		} else if ti.IsReplicaType() {
-			wg.Add(1)
-			go func(i int, ti *topo.TabletInfo) {
-				defer wg.Done()
-				status, err := wr.tmc.ReplicationStatus(ctx, ti.Tablet)
-				if err != nil {
-					rec.RecordError(fmt.Errorf("ReplicationStatus(%v) failed: %v", ti.AliasString(), err))
-					return
-				}
-				result[i] = status
-			}(i, ti)
-		}
-	}
-	wg.Wait()
-	return tablets, result, rec.Error()
-}
 
 // ReparentTablet tells a tablet to reparent this tablet to the current
 // primary, based on the current replication position. If there is no
